chore: clenaup unused window.stuff
